requirements.txt
llama-cpp-python
fastapi
uvicorn
langchain
chromadb
tiktoken
requests
beautifulsoup4
lxml
openai